CAR-TP3/src/test/java/com/lightbend/akka/sample/BodyTest.java
package com.lightbend.akka.sample;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNull;

import org.junit.Before;
import org.junit.Test;

import com.lightbend.akka.sample.MessageFormats.ConnectionMessage;
import com.lightbend.akka.sample.Structure.Body;

import akka.actor.ActorSystem;
import akka.testkit.TestKit;
import scala.concurrent.duration.Duration;
import scala.concurrent.duration.FiniteDuration;

public class BodyTest {

	Body body;
	FiniteDuration fi = Duration.create(15, "seconds");
	
	@Before
	public void initiliaze() {
		this.body = new Body(6);
	}
	
	@Test
	public void generateSystems() {
		this.body.generateSystemes("Test", 1);
		assertEquals(this.body.getSysteme(0).toString(), "Test1");
		assertNull(this.body.getSysteme(1));
	}
	
	
	//Failed tests:   connectionTest(com.lightbend.akka.sample.BodyTest): assertion failed: timeout (3 seconds) during expectMsgClass waiting for class com.lightbend.akka.sample.MessageFormats.ConnectionMessage
	/*
	@Test
	public void connectionTest() {
		this.body.generateSystemes("Systeme", 1);
		final TestKit testKit = new TestKit(this.body.getSysteme(0).getSysteme());
		this.body.generateNoeuds("Noeud", 0, 5, 0);
		this.body.connectionParentChildren(0, 1);
		testKit.expectMsg(fi, "Connexion avec le noeud : Noeud2");
		//ConnectionMessage cm = testKit.expectMsgClass(ConnectionMessage.class);
		//assertEquals("Connexion avec le noeud : Noeud2", cm.toString());
	}*/

}
